gpu-data-webgpu and gpu-struct
